+++ convert/common.c
@@ -261,14 +261,10 @@ static int setup_temp_root_tree(int fd, struct btrfs_mkfs_config *cfg,
 				"extent < dev %llu < %llu, "
 				"dev < fs %llu < %llu, "
 				"fs < csum %llu < %llu",
-				(unsigned long long)root_bytenr,
-				(unsigned long long)extent_bytenr,
-				(unsigned long long)extent_bytenr,
-				(unsigned long long)dev_bytenr,
-				(unsigned long long)dev_bytenr,
-				(unsigned long long)fs_bytenr,
-				(unsigned long long)fs_bytenr,
-				(unsigned long long)csum_bytenr);
+				root_bytenr, extent_bytenr,
+				extent_bytenr, dev_bytenr,
+				dev_bytenr, fs_bytenr,
+				fs_bytenr, csum_bytenr);
 		return -EINVAL;
 	}
 	buf = malloc(sizeof(*buf) + cfg->nodesize);
@@ -433,8 +429,7 @@ static int setup_temp_chunk_tree(int fd, struct btrfs_mkfs_config *cfg,
 	/* Must ensure SYS chunk starts before META chunk */
 	if (meta_chunk_start < sys_chunk_start) {
 		error("wrong chunk order: meta < system %llu < %llu",
-				(unsigned long long)meta_chunk_start,
-				(unsigned long long)sys_chunk_start);
+				meta_chunk_start, sys_chunk_start);
 		return -EINVAL;
 	}
 	buf = malloc(sizeof(*buf) + cfg->nodesize);
@@ -504,8 +499,7 @@ static int setup_temp_dev_tree(int fd, struct btrfs_mkfs_config *cfg,
 	/* Must ensure SYS chunk starts before META chunk */
 	if (meta_chunk_start < sys_chunk_start) {
 		error("wrong chunk order: meta < system %llu < %llu",
-				(unsigned long long)meta_chunk_start,
-				(unsigned long long)sys_chunk_start);
+				meta_chunk_start, sys_chunk_start);
 		return -EINVAL;
 	}
 	buf = malloc(sizeof(*buf) + cfg->nodesize);
@@ -712,16 +706,11 @@ static int setup_temp_extent_tree(int fd, struct btrfs_mkfs_config *cfg,
 				"extent < dev %llu < %llu, "
 				"dev < fs %llu < %llu, "
 				"fs < csum %llu < %llu",
-				(unsigned long long)chunk_bytenr,
-				(unsigned long long)root_bytenr,
-				(unsigned long long)root_bytenr,
-				(unsigned long long)extent_bytenr,
-				(unsigned long long)extent_bytenr,
-				(unsigned long long)dev_bytenr,
-				(unsigned long long)dev_bytenr,
-				(unsigned long long)fs_bytenr,
-				(unsigned long long)fs_bytenr,
-				(unsigned long long)csum_bytenr);
+				chunk_bytenr, root_bytenr,
+				root_bytenr, extent_bytenr,
+				extent_bytenr, dev_bytenr,
+				dev_bytenr, fs_bytenr,
+				fs_bytenr, csum_bytenr);
 		return -EINVAL;
 	}
 	buf = malloc(sizeof(*buf) + cfg->nodesize);
