+++ mkfs/rootdir.c
@@ -343,8 +343,7 @@ static int add_file_items(struct btrfs_trans_handle *trans,
 		ret_read = pread64(fd, buffer, st->st_size, bytes_read);
 		if (ret_read == -1) {
 			error("cannot read %s at offset %llu length %llu: %m",
-				path_name, (unsigned long long)bytes_read,
-				(unsigned long long)st->st_size);
+				path_name, bytes_read, (unsigned long long)st->st_size);
 			free(buffer);
 			goto end;
 		}
@@ -390,10 +389,8 @@ again:
 		ret_read = pread64(fd, eb->data, sectorsize, file_pos +
 				   bytes_read);
 		if (ret_read == -1) {
-			error("cannot read %s at offset %llu length %llu: %m",
-				path_name,
-				(unsigned long long)file_pos + bytes_read,
-				(unsigned long long)sectorsize);
+			error("cannot read %s at offset %llu length %u: %m",
+				path_name, file_pos + bytes_read, sectorsize);
 			goto end;
 		}
